+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.6.0...HEAD) - TBD
+## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.6.1...HEAD) - TBD
+
+## [4.6.1](https://github.com/schemathesis/schemathesis/compare/v4.6.0...v4.6.1) - 2025-11-23
 
 ### :bug: Fixed